Add files via upload
'd like to share my personal project, Production Data Analysis of a Mining Truck with Python. This analysis helps detect quality and machine productivity in seconds, which saves the user significant time while generating reports. The measuring unit used is the Imperial Unit of the United States, but it can easily be modified to the Metric Unit.